+++ ReleaseNotes.md
@@ -6,6 +6,8 @@
 * Badges are now enabled by default
 * SVG badges have been removed in favor of using shields.io via json files from the feed
 * Added prune by release labels option to package retention
+* Added package icon support. Icons will be added to flatcontainer
+* Removed iconUrl support
 
 ## 3.2.1
 * Added badge json for shields.io support [PR](https://github.com/emgarten/Sleet/pull/133)